requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.8.2
    # via flask
cachecontrol==0.14.0
    # via firebase-admin
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
cryptography==43.0.1
    # via pyjwt
firebase-admin==6.5.0
    # via -r requirements.in
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-login
flask-login==0.6.3
    # via -r requirements.in
google-api-core[grpc]==2.20.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.148.0
    # via firebase-admin
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-core==2.4.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.19.0
    # via firebase-admin
google-cloud-storage==2.18.2
    # via firebase-admin
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.66.2
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.66.2
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
msgpack==1.1.0
    # via cachecontrol
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==5.28.2
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.9.0
    # via firebase-admin
pyparsing==3.1.4
    # via httplib2
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
werkzeug==3.0.4
    # via
    #   flask
    #   flask-login
